pyproject.toml
[project]
name = "dz-scrap-api"
version = "0.1.0"
description = "A platform to extract, structure, and classify legal texts from the Official Gazette of Algeria."
authors = []
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "fastapi (>=0.115.13,<0.116.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "numpy (>=2.3.0,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "uvicorn[standard] (>=0.34.3,<0.35.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
